.history/requirements_20250706011437.txt
pandas
numpy
python-dotenv
pymongo
certifi


-e .